chore: Add license for Rhino
Since we are distributing the JAR as part of the source (i.e. it's
checked into the repo) and not the package itself (i.e. it's not
uploaded to http://npmjs.com/ or other registries), we should only need
to distribute the license for Rhino along with the source.
I am not a lawyer so there _might_ be some detail I missed here though.